@@ -3,6 +3,4 @@ const CONNECT_URL = 'https://connect.vezgo.com';
 const API_VERSION = 'v1';
 const API_URL = `${BASE_API_URL}/${API_VERSION}`;
 
-module.exports = {
-  BASE_API_URL, CONNECT_URL, API_VERSION, API_URL,
-};
+export { BASE_API_URL, CONNECT_URL, API_VERSION, API_URL };
